pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cukereuse"
version = "0.1.0"
description = "Static, paraphrase-robust duplicate step detection for Cucumber/Gherkin."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10,<3.13"
authors = [{ name = "Ali Hassaan Mughal" }]
keywords = ["cucumber", "gherkin", "bdd", "duplicate-detection", "clone-detection", "nlp"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "typer>=0.15,<1.0",
  "gherkin-official>=29,<30",
  "sentence-transformers>=3,<4",
  "torch>=2.1,<3.0",
  "python-Levenshtein>=0.25,<0.26",
  "numpy>=1.26,<2.0",
  "scikit-learn>=1.4,<2.0",
  "pandas>=2.2,<3.0",
  "pyarrow>=16,<17",
  "rich>=13,<14",
  "pydantic>=2,<3",
]

[project.optional-dependencies]
mining = [
  "google-cloud-bigquery>=3.17,<4.0",
  "requests>=2.31,<3.0",
]

[project.urls]
Homepage = "https://github.com/amughalbscs16/cukereuse"
Repository = "https://github.com/amughalbscs16/cukereuse"
Issues = "https://github.com/amughalbscs16/cukereuse/issues"

[project.scripts]
cukereuse = "cukereuse.cli:app"

[dependency-groups]
dev = [
  "pytest>=8,<9",
  "pytest-cov>=5,<6",
  "ruff>=0.5,<0.6",
  "mypy>=1.10,<2.0",
  "matplotlib>=3.10.8",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]

[tool.hatch.build.targets.wheel]
packages = ["src/cukereuse"]

[tool.hatch.build.targets.sdist]
include = ["src/cukereuse", "README.md", "LICENSE", "CITATION.cff", "pyproject.toml"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
pythonpath = ["src"]

[tool.coverage.run]
source = ["cukereuse"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
  "E", "F", "W",   # pycodestyle + pyflakes
  "I",             # isort
  "N",             # pep8-naming
  "UP",            # pyupgrade
  "B",             # bugbear
  "C4",            # comprehensions
  "SIM",           # simplify
  "TCH",           # type-checking imports
  "RUF",           # ruff-specific
]
ignore = ["E501"]  # line too long handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N802"]

[tool.mypy]
python_version = "3.10"
strict = true
files = ["src/cukereuse"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*", "scripts.*"]
strict = false
